Move the price rounding tests to the org test package
apply_price_rounding lives in onegov.org.utils, so testing it from
tests.onegov.pay made the lower layer import the higher one.
round_to_five_rappen stays with onegov.pay.
